requirements.txt
torch
torchaudio
deepfilternet
matplotlib
gradio
Pillow
django
djangorestframework
pydub
